fixed conversion issue in Game::SpawnAliens()

+++ Source/game.cpp
@@ -464,8 +464,8 @@ void Game::SpawnAliens()
 		for (int col = 0; col < formationWidth; col++) {
 			auto newAlien = Alien();
 			newAlien.active = true;
-			newAlien.position.x = formationX + 450 + (col * alienSpacing);
-			newAlien.position.y = formationY + (row * alienSpacing);
+			newAlien.position.x = (float)formationX + 450 + ((float)col * (float)alienSpacing);
+			newAlien.position.y = (float)formationY + ((float)row * (float)alienSpacing);
 			Aliens.push_back(newAlien);
 			std::cout << "Find Alien -X:" << newAlien.position.x << std::endl;
 			std::cout << "Find Alien -Y:" << newAlien.position.y << std::endl;
